Move plan into crate::common, simplify interpreter module
