+++ lib/src/common/plan.rs
