+++ lib/src/metta/interpreter.rs
@@ -1,5 +1,5 @@
-use super::*;
 use crate::*;
+use crate::common::plan::*;
 use crate::atom::subexpr::*;
 use crate::atom::matcher::*;
 use crate::space::grounding::*;
