--- lib/src/metta/interpreter/mod.rs
@@ -1,5 +1,0 @@
-mod plan;
-mod matching;
-
-pub use plan::*;
-pub use matching::interpret;
